+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fixed the way `internet_accessible` bool is referenced
 - added `manifests/local` for local testing and `examples/manifests/example` as a guidance for calling the modules using `git paths`
 - added contributing guidance to `CONTRIBUTING.md`
+- refactored AWS Batch module, to be agnostic of a project with unit-tests
 
 ### **Changed**